changed smp unet w Resnext

+++ lit_models/UNET_monai_lit.py
@@ -180,7 +180,7 @@ class UNET_lit(pl.LightningModule):
         #return self.loss_focal(y_pred * mask, y_true) + self.loss_tversky(y_pred * mask, y_true)
         #return self.monai_masked_tversky(y_pred, y_true, mask) +  self.masked_focal(y_pred, y_true, mask)
 
-        return self.monai_masked_tversky(y_pred, y_true, mask) +  self.loss_bce(y_pred*mask, y_true)
+        return self.monai_masked_tversky(y_pred, y_true, mask) +  self.loss_bce(y_pred*mask, y_true.float())
 
 
     def combined_loss(self, pred, label, mask):
